Reduce the amount of duplicate test cases. As these evolved some of the more complicated test cases encompassed some of the original simpler ones
