+++ src/test/kotlin/com/cjbooms/fabrikt/generators/ModelGeneratorTest.kt
@@ -19,37 +19,24 @@ import org.junit.jupiter.params.provider.MethodSource
 class ModelGeneratorTest {
 
     private fun testCases(): Stream<String> = Stream.of(
-        "optionalVsRequired",
-        "snakeToCamel",
-        "arraySimpleInLined",
-        "objectRef",
-        "inLinedObject",
-        "arrayRef",
-        "simpleRef",
-        "allOf",
-        "anyOf",
-        "requiredReadOnly",
-        "polymorphicModels",
-        "oneOfPolymorphicModels",
-        "validationAnnotations",
-        "wildCardTypes",
-        "nullableListTypes",
-        "oneOfAndTopLevelProps",
-        "oneOf",
-        "complexAnyOneAllCombo",
-        "duplicatePropertyHandling",
-        "allOfNestedOneOf",
-        "deepNestedReferences",
-        "sharingObjectDefinitions",
+        "arrays",
+        "anyOfOneOfAllOf",
         "deepNestedSharingReferences",
-        "mixingCamelSnakeLispCase",
-        "mapExamples",
-        "enumExamples",
-        "arrayOfArrays",
         "defaultValues",
+        "duplicatePropertyHandling",
+        "enumExamples",
         "enumPolymorphicDiscriminator",
         "externalReferences",
-        "githubApi"
+        "githubApi",
+        "inLinedObject",
+        "mapExamples",
+        "mixingCamelSnakeLispCase",
+        "oneOfPolymorphicModels",
+        "optionalVsRequired",
+        "polymorphicModels",
+        "requiredReadOnly",
+        "validationAnnotations",
+        "wildCardTypes"
     )
 
     @ParameterizedTest
