--- src/test/resources/examples/allOf/api.yaml
@@ -1,32 +1,0 @@
-openapi: 3.0.0
-components:
-  schemas:
-    ContainsAllOf:
-      allOf:
-        - $ref: "#/components/schemas/SomeAllOf"
-        - type: object
-          required:
-            - weight_on_mars
-          properties:
-            weight_on_mars:
-              type: integer
-    SomeAllOf:
-      type: object
-      required:
-        - prop_one
-        - prop_two
-      properties:
-        prop_one:
-          type: string
-        prop_two:
-          type: string
-          format: date-time
-        primitive_allOf:
-          description: "Some good description"
-          allOf:
-            - $ref: "#/components/schemas/PrimitiveInt"
-
-    PrimitiveInt:
-      type: integer
-      maximum: 5
-      minimum: 1
