--- src/test/resources/examples/anyOf/api.yaml
@@ -1,23 +1,0 @@
-openapi: 3.0.0
-components:
-  schemas:
-    ContainsAnyOf:
-      anyOf:
-        - $ref: "#/components/schemas/SomeAnyOf"
-        - type: object
-          required:
-            - weight_on_mars
-          properties:
-            weight_on_mars:
-              type: integer
-    SomeAnyOf:
-      type: object
-      required:
-        - prop_one
-        - prop_two
-      properties:
-        prop_one:
-          type: string
-        prop_two:
-          type: string
-          format: date-time
