+++ src/test/resources/examples/anyOfOneOfAllOf/api.yaml
@@ -47,3 +47,32 @@ components:
       properties:
         second_nested_any_of_prop:
           type: string
+          
+    OneOfAndTopLevelProps:
+      type: object
+      required:
+        - forth_property
+        - fifth_property
+      properties:
+        forth_property:
+          type: string
+        fifth_property:
+          type: string
+          format: date-time
+      oneOf:
+        - $ref: "#/components/schemas/FirstOneB"
+        - $ref: "#/components/schemas/SecondOneB"
+
+    FirstOneB:
+      type: object
+      properties:
+        first_property:
+          type: string
+
+    SecondOneB:
+      type: object
+      properties:
+        second_property:
+          type: string
+        third_property:
+          type: string
