+++ src/test/resources/examples/anyOfOneOfAllOf/models/Models.kt
@@ -1,7 +1,9 @@
-package examples.complexAnyOneAllCombo.models
+package examples.anyOfOneOfAllOf.models
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import java.math.BigDecimal
+import java.time.OffsetDateTime
+import javax.validation.constraints.NotNull
 import kotlin.String
 
 data class ComplexParent(
@@ -68,4 +70,39 @@ data class SecondAnyA(
     @param:JsonProperty("second_nested_any_of_prop")
     @get:JsonProperty("second_nested_any_of_prop")
     val secondNestedAnyOfProp: String? = null
+)
+
+data class OneOfAndTopLevelProps(
+    @param:JsonProperty("first_property")
+    @get:JsonProperty("first_property")
+    val firstProperty: String? = null,
+    @param:JsonProperty("second_property")
+    @get:JsonProperty("second_property")
+    val secondProperty: String? = null,
+    @param:JsonProperty("third_property")
+    @get:JsonProperty("third_property")
+    val thirdProperty: String? = null,
+    @param:JsonProperty("forth_property")
+    @get:JsonProperty("forth_property")
+    @get:NotNull
+    val forthProperty: String,
+    @param:JsonProperty("fifth_property")
+    @get:JsonProperty("fifth_property")
+    @get:NotNull
+    val fifthProperty: OffsetDateTime
+)
+
+data class FirstOneB(
+    @param:JsonProperty("first_property")
+    @get:JsonProperty("first_property")
+    val firstProperty: String? = null
+)
+
+data class SecondOneB(
+    @param:JsonProperty("second_property")
+    @get:JsonProperty("second_property")
+    val secondProperty: String? = null,
+    @param:JsonProperty("third_property")
+    @get:JsonProperty("third_property")
+    val thirdProperty: String? = null
 )
