--- src/test/resources/examples/arrayOfArrays/api.yaml
@@ -1,26 +1,0 @@
-openapi: 3.0.0
-info:
-paths:
-components:
-  schemas:
-    ContainsArrayOfArrays:
-      type: object
-      properties:
-        array_of_arrays:
-          type: array
-          items:
-            $ref: "#/components/schemas/ArrayOfSomething"
-
-    ArrayOfSomething:
-      type: array
-      items:
-        $ref: "#/components/schemas/Something"
-
-    Something:
-      type: object
-      required:
-        grams
-      properties:
-        some_value:
-          type: integer
-          example: 100
