--- src/test/resources/examples/arraySimpleInLined/api.yaml
@@ -1,13 +1,0 @@
-openapi: 3.0.0
-components:
-  schemas:
-    ArraySimpleInLined:
-      type: object
-      required:
-        - quantities
-      properties:
-        quantities:
-          type: array
-          items:
-            type: integer
-            format: int64
