--- src/test/resources/examples/arraySimpleInLined/models/Models.kt
@@ -1,13 +1,0 @@
-package examples.arraySimpleInLined.models
-
-import com.fasterxml.jackson.annotation.JsonProperty
-import javax.validation.constraints.NotNull
-import kotlin.Long
-import kotlin.collections.List
-
-data class ArraySimpleInLined(
-    @param:JsonProperty("quantities")
-    @get:JsonProperty("quantities")
-    @get:NotNull
-    val quantities: List<Long>
-)
