--- src/test/resources/examples/deepNestedReferences/api.yaml
@@ -1,43 +1,0 @@
-openapi: 3.0.0
-components:
-  schemas:
-    ContainsDeepReferences:
-      type: object
-      properties:
-        weight_on_mars:
-          type: array
-          items:
-            $ref: "#/components/schemas/FirstLevelArrayRef"
-
-    FirstLevelArrayRef:
-      type: object
-      properties:
-        gravity:
-          type: object
-          $ref: "#/components/schemas/SecondLevelObjectRef"
-        weights:
-          type: array
-          items:
-            $ref: "#/components/schemas/SecondLevelArrayRef"
-
-    SecondLevelObjectRef:
-      type: object
-      properties:
-        gravity:
-          type: integer
-          example: 99
-        third_level:
-          $ref: "#/components/schemas/ThirdLevelObject"
-
-    SecondLevelArrayRef:
-      type: object
-      properties:
-        grams:
-          type: integer
-          example: 100
-
-    ThirdLevelObject:
-      type: object
-      properties:
-        third:
-          type: string
