--- src/test/resources/examples/nullableListTypes/api.yaml
@@ -1,14 +1,0 @@
-openapi: 3.0.0
-components:
-  schemas:
-    NullableListTypes:
-      type: object
-      required:
-        - name
-      properties:
-        name:
-          type: string
-        tags:
-          type: array
-          items:
-            type: string
