--- src/test/resources/examples/objectRef/api.yaml
@@ -1,18 +1,0 @@
-openapi: 3.0.0
-components:
-  schemas:
-    ContainsObjectRef:
-      type: object
-      required:
-        - weight_on_mars
-      properties:
-        weight_on_mars:
-          $ref: "#/components/schemas/ObjectRef"
-    ObjectRef:
-      type: object
-      required:
-        grams
-      properties:
-        grams:
-          type: integer
-          example: 100
