--- src/test/resources/examples/oneOfAndTopLevelProps/api.yaml
@@ -1,31 +1,0 @@
-openapi: 3.0.0
-components:
-  schemas:
-    OneOfAndTopLevelProps:
-      type: object
-      required:
-        - forth_property
-        - fifth_property
-      properties:
-        forth_property:
-          type: string
-        fifth_property:
-          type: string
-          format: date-time
-      oneOf:
-        - $ref: "#/components/schemas/FirstOneB"
-        - $ref: "#/components/schemas/SecondOneB"
-
-    FirstOneB:
-      type: object
-      properties:
-        first_property:
-          type: string
-
-    SecondOneB:
-      type: object
-      properties:
-        second_property:
-          type: string
-        third_property:
-          type: string
