--- src/test/resources/examples/putApi/service/Services.kt
@@ -1,12 +1,0 @@
-package examples.putApi.service
-
-import examples.putApi.models.Contributor
-import kotlin.String
-
-interface ContributorsService {
-    fun update(
-        contributor: Contributor,
-        id: String,
-        ifMatch: String
-    ): Contributor
-}
