--- src/test/resources/examples/sharingObjectDefinitions/api.yaml
@@ -1,23 +1,0 @@
-openapi: 3.0.0
-components:
-  schemas:
-    FirstContainingShared:
-      type: object
-      properties:
-        weight_on_mars:
-          $ref: "#/components/schemas/SharedDefinition"
-    SecondContainingShared:
-      type: object
-      required:
-        - weight_on_mars
-      properties:
-        weight_on_mars:
-          $ref: "#/components/schemas/SharedDefinition"
-    SharedDefinition:
-      type: object
-      required:
-        grams
-      properties:
-        grams:
-          type: integer
-          example: 100
