--- src/test/resources/examples/simpleRequestBody/service/Services.kt
@@ -1,9 +1,0 @@
-package examples.simpleRequestBody.service
-
-import java.net.URI
-import kotlin.Pair
-import kotlin.String
-
-interface UploadService {
-    fun create(oASDocument: String): Pair<URI, String?>
-}
