--- src/test/resources/examples/snakeToCamel/api.yaml
@@ -1,10 +1,0 @@
-openapi: 3.0.0
-components:
-  schemas:
-    SnakeToCamel:
-      type: object
-      required:
-        - user_name
-      properties:
-        user_name:
-          type: string
